backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pillow
pytesseract
requests
google-generativeai
python-dotenv
